+++ Release Notes.md
@@ -5,6 +5,7 @@
 
 ### ✨ New features
 
+* `MenuListItem` now lets you define a title view in a block.
 * `UserDefaultsPersisted` now lets you inject a custom store.